agentBuiltApps/blogPostMaker/requirements.txt
# Web framework
flask>=2.3.0
requests>=2.31.0

# AI/ML libraries
openai>=1.0.0
semantic-kernel>=0.9.0

# Text processing
markdown>=3.5.0
beautifulsoup4>=4.12.0

# Configuration
python-dotenv>=1.0.0

# Additional utilities
Pillow>=10.0.0
jinja2>=3.1.0

google-generativeai>=0.3.0